requirements.txt
pysensors
liquidctl